cookbooks/python/composio/requirements.txt
swarmkit
composio
python-dotenv
rich
prompt_toolkit
